@@ -1,15 +1,18 @@
 <script>
-  import { Meta, Story } from '@storybook/addon-svelte-csf'
-  import Embed from '../../src/modules/embed'
+	import { Meta, Story } from '@storybook/addon-svelte-csf';
+	import Embed from '../../src/lib/modules/embed';
+	import JQuery from '../../src/lib/common/JQuery.svelte';
 
-  /**
-   * @type {Embed}
-   */
-  let controller
+	/**
+	 * @type {Embed}
+	 */
+	let controller;
 </script>
 
-<Meta title="Modules/Embed" component="{Embed}" />
+<Meta title="Modules/Embed" component={Embed} />
 
 <Story name="Solo" let:args>
-  <Embed {...args} bind:this="{controller}" />
+	<JQuery>
+		<Embed {...args} bind:this={controller} />
+	</JQuery>
 </Story>
