@@ -1,43 +1,39 @@
 <script>
-  import { Meta, Template, Story } from '@storybook/addon-svelte-csf'
-  import { Centerize } from '../components'
-  import * as Modal from '../../src/modules/modal'
-  import Button from '../../src/elements/button'
-  import Icon from '../../src/elements/icon'
+	import { Meta, Template, Story } from '@storybook/addon-svelte-csf';
+	import { Centerize } from '../components';
+	import * as Modal from '../../src/lib/modules/modal';
+	import Button from '../../src/lib/elements/button';
+	import Icon from '../../src/lib/elements/icon';
+	import JQuery from '../../src/lib/common/JQuery.svelte';
 
-  /**
-   * @type {Modal.default}
-   */
-  let controller
+	/**
+	 * @type {Modal.default}
+	 */
+	let controller;
 </script>
 
-<Meta title="Modules/Modal" component="{Modal.default}" />
+<Meta title="Modules/Modal" component={Modal.default} />
 
-<Centerize>
-  <Template let:args>
-    <Modal.default {...args} bind:this="{controller}">
-      <Modal.Header icon>
-        <Icon name="close" />
-        Archive Old Messages
-      </Modal.Header>
-      <Modal.Content>
-        <p>
-          Your inbox is getting full, would you like us to enable automatic
-          archiving of old messages?
-        </p>
-      </Modal.Content>
-      <Modal.Actions>
-        <Button>No</Button>
-        <Button>Yes</Button>
-      </Modal.Actions>
-    </Modal.default>
+<Story name="Primary" let:args>
+	<Modal.default {...args} bind:this={controller}>
+		<Modal.Header icon>
+			<Icon name="close" />
+			Archive Old Messages
+		</Modal.Header>
+		<Modal.Content>
+			<p>
+				Your inbox is getting full, would you like us to enable automatic archiving of old messages?
+			</p>
+		</Modal.Content>
+		<Modal.Actions>
+			<Button>No</Button>
+			<Button>Yes</Button>
+		</Modal.Actions>
+	</Modal.default>
 
-    <Button
-      on:click="{() => {
-        controller.show()
-      }}">Open Modal</Button
-    >
-  </Template>
-</Centerize>
-
-<Story name="Primary" />
+	<Button
+		on:click={() => {
+			controller.show();
+		}}>Open Modal</Button
+	>
+</Story>
