@@ -1,8 +1,9 @@
 <script>
   import { Meta, Story } from '@storybook/addon-svelte-csf'
-  import Popup, { Header, Body } from '../../src/modules/popup'
-  import Button from '../../src/elements/button'
+  import Popup, { Header, Body } from '../../src/lib/modules/popup'
+  import Button from '../../src/lib/elements/button'
   import { Centerize } from '../components'
+	import JQuery from '../../src/lib/common/JQuery.svelte';
 
   /**
    * @type {Popup}
@@ -14,12 +15,14 @@
 
 <Centerize>
   <Story name="Solo" let:args>
-    <Popup {...args} let:popup>
-      <Button onMount="{popup}">Click me</Button>
-      <Body>
-        <Header>Notice</Header>
-        This is just a random text !
-      </Body>
-    </Popup>
+    <JQuery>
+      <Popup {...args} let:popup>
+        <Button onMount="{popup}">Click me</Button>
+        <Body>
+          <Header>Notice</Header>
+          This is just a random text !
+        </Body>
+      </Popup>
+    </JQuery>
   </Story>
 </Centerize>
