@@ -1,7 +1,8 @@
 <script>
     import { Meta, Story } from '@storybook/addon-svelte-csf'
-    import Rating from '../../src/modules/rating'
-  
+    import Rating from '../../src/lib/modules/rating'
+  	import JQuery from '../../src/lib/common/JQuery.svelte';
+
     /**
      * @type {Rating}
      */
@@ -11,6 +12,8 @@
   <Meta title="Modules/Rating" component="{Rating}" />
   
   <Story name="Solo" let:args>
-    <Rating {...args} bind:this="{controller}" />
+    <JQuery>
+      <Rating {...args} bind:this="{controller}" />
+    </JQuery>
   </Story>
   
