@@ -1,7 +1,8 @@
 <script>
     import { Meta, Story } from '@storybook/addon-svelte-csf'
-    import Search from '../../src/modules/search'
-  
+    import Search from '../../src/lib/modules/search'
+  	import JQuery from '../../src/lib/common/JQuery.svelte';
+
     /**
      * @type {Search}
      */
@@ -11,6 +12,8 @@
   <Meta title="Modules/Search" component="{Search}" />
   
   <Story name="Solo" let:args>
-    <Search {...args} bind:this="{controller}" />
+    <JQuery>
+      <Search {...args} bind:this="{controller}" />
+    </JQuery>
   </Story>
   
