@@ -1,6 +1,7 @@
 <script>
   import { Meta, Story } from '@storybook/addon-svelte-csf'
-  import Slider from '../../src/modules/slider'
+  import Slider from '../../src/lib/modules/slider'
+	import JQuery from '../../src/lib/common/JQuery.svelte';
 
   /**
    * @type {Slider}
@@ -11,5 +12,7 @@
 <Meta title="Modules/Slider" component="{Slider}" />
 
 <Story name="Solo" let:args>
-  <Slider {...args} bind:this="{controller}" />
+  <JQuery>
+    <Slider {...args} bind:this="{controller}" />
+  </JQuery>
 </Story>
