@@ -1,7 +1,8 @@
 <script>
     import { Meta, Story } from '@storybook/addon-svelte-csf'
-    import Sticky from '../../src/modules/sticky'
-  
+    import Sticky from '../../src/lib/modules/sticky'
+  	import JQuery from '../../src/lib/common/JQuery.svelte';
+
     /**
      * @type {Sticky}
      */
@@ -11,6 +12,9 @@
   <Meta title="Modules/Sticky" component="{Sticky}" />
   
   <Story name="Solo" let:args>
-    <Sticky {...args} bind:this="{controller}" />
+    <JQuery>
+      <Sticky {...args} bind:this="{controller}" />
+    
+    </JQuery>
   </Story>
   
